test(robot): add test case Test Snapshot Purge Rejection While Migration
Reject snapshot purge when volume is being live migration.
If snapshot purge is executed while live migrating a volume,
the disk files will be changed. The new engine and replicas
are unable to be aware of the change and cannot list the new
disk files, resulting in ERR replicas.

Longhorn 10495
Longhorn 10496

Signed-off-by: Derek Su <derek.su@suse.com>